Sep Comms into two; Change getSTDM return type
Create source folder and move the C++ implementation
of the Communications class there.

Change the return type of getSTDM method from a struct to
a reference.

Include the ostream library into SPLStandarMessage.h as
it is required for ostream operations.
