+++ edante/comms/test/Makefile
@@ -1,6 +1,7 @@
 CXXFLAGS = -std=c++11 -I../include
 
 test-comms:
+	$(CXX) $(CXXFLAGS) test-comms.cpp ../src/Communication.cpp -o test-comms
 
 clean:
 	rm test-comms
